backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
scikit-learn>=1.4.0
lifelines>=0.27.0
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
hypothesis>=6.100.0
lxml>=5.0.0
